AXON-624: moved helper to the common.ts
